+++ datapipe/cli.py
@@ -168,7 +168,14 @@ def cli(
         trace.get_tracer_provider().add_span_processor(span_processor)  # type: ignore
 
     if trace_gcp:
-        from opentelemetry.exporter.cloud_trace import CloudTraceSpanExporter
+        try:
+            from opentelemetry.exporter.cloud_trace import (  # type: ignore
+                CloudTraceSpanExporter,
+            )
+        except ImportError:
+            raise ImportError(
+                "Please install opentelemetry-exporter-cloud-trace to use GCP Trace"
+            )
 
         cloud_trace_exporter = CloudTraceSpanExporter(
             resource_regex=r".*",
@@ -200,18 +207,18 @@ def table():
     pass
 
 
-@table.command()
+@table.command(name="list")
 @click.pass_context
-def list(ctx: click.Context) -> None:
+def table_list(ctx: click.Context) -> None:
     app: DatapipeApp = ctx.obj["pipeline"]
 
     for table in sorted(app.catalog.catalog.keys()):
         print(table)
 
 
-@cli.command()
+@cli.command(name="run")
 @click.pass_context
-def run(ctx: click.Context) -> None:
+def main_run(ctx: click.Context) -> None:
     app: DatapipeApp = ctx.obj["pipeline"]
 
     with tracer.start_as_current_span("run"):
@@ -325,7 +332,7 @@ def to_human_repr(step: ComputeStep, extra_args: Optional[Dict] = None) -> str:
         labels = " ".join([f"[magenta]{k}={v}[/magenta]" for (k, v) in step.labels])
         res.append(f"  labels: {labels}")
 
-    if inputs_arr := [i.name for i in step.input_dts]:
+    if inputs_arr := [inp.dt.name for inp in step.input_dts]:
         inputs = ", ".join(inputs_arr)
         res.append(f"  inputs: {inputs}")
 
@@ -340,10 +347,10 @@ def to_human_repr(step: ComputeStep, extra_args: Optional[Dict] = None) -> str:
     return "\n".join(res)
 
 
-@step.command()  # type: ignore
+@step.command(name="list")  # type: ignore
 @click.option("--status", is_flag=True, type=click.BOOL, default=False)
 @click.pass_context
-def list(ctx: click.Context, status: bool) -> None:  # noqa
+def step_list(ctx: click.Context, status: bool) -> None:  # noqa
     app: DatapipeApp = ctx.obj["pipeline"]
     steps: List[ComputeStep] = ctx.obj["steps"]
 
@@ -366,13 +373,13 @@ def list(ctx: click.Context, status: bool) -> None:  # noqa
         rprint("")
 
 
-@step.command()  # type: ignore
+@step.command(name="run")
 @click.option("--loop", is_flag=True, default=False, help="Run continuosly in a loop")
 @click.option(
     "--loop-delay", type=click.INT, default=30, help="Delay between loops in seconds"
 )
 @click.pass_context
-def run(ctx: click.Context, loop: bool, loop_delay: int) -> None:  # noqa
+def step_run(ctx: click.Context, loop: bool, loop_delay: int) -> None:
     app: DatapipeApp = ctx.obj["pipeline"]
     steps_to_run: List[ComputeStep] = ctx.obj["steps"]
 
@@ -393,7 +400,7 @@ def run(ctx: click.Context, loop: bool, loop_delay: int) -> None:  # noqa
             print("\n\n")
 
 
-@step.command()  # type: ignore
+@step.command()
 @click.argument("idx", type=click.STRING)
 @click.pass_context
 def run_idx(ctx: click.Context, idx: str) -> None:
@@ -409,7 +416,7 @@ def run_idx(ctx: click.Context, idx: str) -> None:
             step.run_idx(ds=app.ds, idx=cast(IndexDF, pd.DataFrame([idx_dict])))
 
 
-@step.command()  # type: ignore
+@step.command()
 @click.option("--loop", is_flag=True, default=False, help="Run continuosly in a loop")
 @click.option(
     "--loop-delay", type=click.INT, default=1, help="Delay between loops in seconds"
@@ -489,7 +496,7 @@ def fill_metadata(ctx: click.Context) -> None:
             step.fill_metadata(app.ds)
 
 
-@step.command()  # type: ignore
+@step.command()
 @click.pass_context
 def reset_metadata(ctx: click.Context) -> None:  # noqa
     app: DatapipeApp = ctx.obj["pipeline"]
@@ -516,8 +523,14 @@ def migrate_transform_tables(ctx: click.Context, labels: str, name: str) -> None
     return migrations_v013.migrate_transform_tables(app, batch_transforms_steps)
 
 
-for entry_point in metadata.entry_points().get("datapipe.cli", []):
-    register_commands = entry_point.load()
+try:
+    entry_points = metadata.entry_points(group="datapipe.cli")  # type: ignore
+except TypeError:
+    # Compatibility with older versions of importlib.metadata (Python 3.8-3.9)
+    entry_points = metadata.entry_points().get("datapipe.cli", [])  # type: ignore
+
+for entry_point in entry_points:
+    register_commands = entry_point.load()  # type: ignore
     register_commands(cli)