+++ examples/many_to_zero/app.py
@@ -1,35 +1,13 @@
-from typing import Dict, List, Optional
-
 import pandas as pd
 from sqlalchemy import Integer
 from sqlalchemy.sql.schema import Column
 
 from datapipe.compute import Catalog, DatapipeApp, Pipeline, Table
-from datapipe.datatable import DataStore, DataTable
-from datapipe.run_config import RunConfig
+from datapipe.datatable import DataStore
 from datapipe.step.batch_generate import BatchGenerate
-from datapipe.step.datatable_transform import DatatableTransform
+from datapipe.step.batch_transform import BatchTransform
 from datapipe.store.database import DBConn
 from datapipe.store.pandas import TableStoreJsonLine
-
-catalog = Catalog(
-    {
-        "input": Table(
-            store=TableStoreJsonLine(
-                filename="input.json",
-                primary_schema=[
-                    Column("id", Integer, primary_key=True),
-                ],
-            )
-        ),
-        "result": Table(
-            store=TableStoreJsonLine(
-                filename="result.json",
-                primary_schema=[Column("result_id", Integer, primary_key=True)],
-            )
-        ),
-    }
-)
 
 
 def generate_data():
@@ -42,36 +20,41 @@ def generate_data():
 
 
 def count(
-    ds: DataStore,
-    input_dts: List[DataTable],
-    output_dts: List[DataTable],
-    kwargs: Dict,
-    run_config: Optional[RunConfig] = None,
-) -> None:
-    assert len(input_dts) == 1
-    assert len(output_dts) == 1
+    input_df: pd.DataFrame,
+) -> pd.DataFrame:
+    return pd.DataFrame({"result_id": [0], "count": [len(input_df)]})
 
-    input_dt = input_dts[0]
-    output_dt = output_dts[0]
 
-    output_dt.store_chunk(
-        pd.DataFrame(
-            {"result_id": [0], "count": [len(input_dt.meta_table.get_existing_idx())]}
-        )
-    )
+input_tbl = Table(
+    name="input",
+    store=TableStoreJsonLine(
+        filename="input.json",
+        primary_schema=[
+            Column("id", Integer, primary_key=True),
+        ],
+    ),
+)
+
+result_tbl = Table(
+    name="result",
+    store=TableStoreJsonLine(
+        filename="result.json",
+        primary_schema=[Column("result_id", Integer, primary_key=True)],
+    ),
+)
 
 
 pipeline = Pipeline(
     [
         BatchGenerate(
             generate_data,
-            outputs=["input"],
+            outputs=[input_tbl],
         ),
-        DatatableTransform(
-            count,  # type: ignore
-            inputs=["input"],
-            outputs=["result"],
-            check_for_changes=False,
+        BatchTransform(
+            count,
+            inputs=[input_tbl],
+            outputs=[result_tbl],
+            transform_keys=[],
         ),
     ]
 )
@@ -79,4 +62,10 @@ pipeline = Pipeline(
 
 ds = DataStore(DBConn("sqlite+pysqlite3:///db.sqlite"))
 
-app = DatapipeApp(ds, catalog, pipeline)
+app = DatapipeApp(ds, Catalog({}), pipeline)
+
+
+if __name__ == "__main__":
+    from datapipe.compute import run_steps
+
+    run_steps(ds, app.steps, None, None)