+++ tests/test_batch_transform_scheduling.py
@@ -3,6 +3,7 @@ from typing import List
 import pandas as pd
 from sqlalchemy import Column, Integer
 
+from datapipe.compute import ComputeInput
 from datapipe.datatable import DataStore
 from datapipe.step.batch_transform import BatchTransformStep
 from datapipe.store.database import TableStoreDB
@@ -42,7 +43,9 @@ def test_inc_process_proc_no_change(dbconn) -> None:
         ds=ds,
         name="step",
         func=id_func,
-        input_dts=[tbl1],
+        input_dts=[
+            ComputeInput(dt=tbl1, join_type="full"),
+        ],
         output_dts=[tbl2],
     )
 
@@ -148,7 +151,11 @@ def test_aux_input(dbconn) -> None:
         ds=ds,
         name="step",
         func=lambda items1, items2, aux: items1,
-        input_dts=[tbl_items1, tbl_items2, tbl_aux],
+        input_dts=[
+            ComputeInput(dt=tbl_items1, join_type="full"),
+            ComputeInput(dt=tbl_items2, join_type="full"),
+            ComputeInput(dt=tbl_aux, join_type="full"),
+        ],
         output_dts=[tbl_out],
         transform_keys=["id"],
     )